small tweaks

+++ 09-nizovi-arm64/04-minimax/main.cpp
@@ -1,18 +1,18 @@
 #include <iostream>
 #include <vector>
 
-extern "C" void minimax(int *a, int n, int &pmin, int &pmax);
+extern "C" void minimax(short *a, int n, short &pmin, short &pmax);
 
 int main() {
     int n;
     std::cin >> n;
 
-    std::vector<int> a(n);
+    std::vector<short> a(n);
     for (int i = 0; i < n; i++) {
         std::cin >> a[i];
     }
 
-    int min, max;
+    short min, max;
     minimax(a.data(), n, min, max);
 
     std::cout << min << " " << max << std::endl;
